Issue-114: Renamed package for IReviewListMapper.kt
